planner-logging-refactor
